refactor(MWA): Make API parts persistent
Using `include` re-evaluates modules every single time.
This means that any module-local variable gets discarded when calling
`luamod` and we don't want that. But we *do* want to use `include` for
easy mod reloading.

The API's modules are now stored into their ModReference, which is called
with `require` in order to keep it cached.

The file `mod.lua` must ALWAYS be imported using `require`. The rest calling
be safely called using `include` and should be either initialised if not already
or return the existing instance.