refactor(EntityMelee): Use a proxy for `:GetState()`
Instead of cloning the table when using `:GetState()` in safe mode (the
default), provide user with a read-only proxy to the state table.
This ensures the performance of `:GetState()` will ALWAYS be the same no
matter the size of the state table.

Note: This doesn't prevent non-primitive values from being edited.
*Might* Want to look into that later, but right now that sounds like
overdoing things.