refactor(*): Improve global usage
Some mod(s), at least Regret Pedestals, use `_G.mod` and reloading the
API using `luamod` breaks that—because it deletes `_G.mod`.
This commit fixes that by using the MeleeWeaponApi as index for storing
the `ModReference` into `_G`.
This method also allows access to users of the API who wish to get
access to the internals of the API for whatever reason.
We do a little bit of hacking 😎